+++ Anathema/GUI/Tools/Scanners/GUIInputCorrelator.cs
@@ -45,7 +45,7 @@ namespace Anathema
         private void SetVariableSize()
         {
             Int32 VariableSize = (Int32)Math.Pow(2, VariableSizeTrackBar.Value);
-            VariableSizeValueLabel.Text = Conversions.BytesToMetric((UInt64)VariableSize).ToString();
+            VariableSizeValueLabel.Text = Conversions.BytesToMetric(VariableSize).ToString();
 
             InputCorrelatorPresenter.SetVariableSize(VariableSize);
         }
